Beginning of User Creation and DB Schema
This has a lot of work done in the User class with plenty of future
changes to be made. Also, a DonationCenter class was created (no work
done). More importantly, a DatabaseContext was made that should be
able to create the schema (hasn't been run yet so I don't think there
is anything in the DB yet). Will be adding migrations ability soon so
it can handle changes in code and reflect them in DB